+++ common/tests/sema4ai_common_tests/test_autoexit.py
@@ -1,5 +1,6 @@
 def test_exit_when_pid_exists():
     import sys
+    import time
 
     from sema4ai.common.process import Process
     from sema4ai.common.wait_for import wait_for_condition
@@ -27,6 +28,11 @@ def test_exit_when_pid_exists():
             timeout=15,
             msg=lambda: f"Watcher process was not created in time. Output:\n{stream.getvalue()}",
         )
+
+        # Just wait a bit to make sure that we don't kill before the pid to track is killed.
+        time.sleep(0.2)
+        assert target_process.is_alive()
+        assert watcher_process.is_alive()
 
         # Kill the target process
         target_process.stop()